Implement getters and setters for NGCHM types, chmLoadShaidyCHM
Getters and setters (e.g. chmProperty, chmCovariateBar) allow the
user to upload some properties without directly referencing the object's
slots.

Initial, partial implementation of chmLoadShaidyCHM. It allows user to
download an ngchm object from a shaidy server.  The goal is to allow
the R user to programmaticaly download an ngchm object, modify it, and
upload the modified ngchm to a possibly different server.  E.g. add your
own covariate of interest to a map in our TCGA compendium.