Do not expose TreeCut in chmNew
In the chmNew function 'rowTreeGaps' and 'colTreeGaps' are no longer
expposed to the user. Instead, only allow the user to specify
'rowGapLocations' and  'colGapLocations'. The user would specify tree
gaps with, for example:
  chmNew("ngchm", rowGapLocations=chmTreeCuts(3))
which sets the chm slot 'rowTreeCuts' to 3.

This is accomplished in the initialization of the chm object. If the
class of 'rowGapLocations' is 'treeCuts' (as in the above example), then
the 'rowTreeCuts' slot is set to the given value, and 'rowCutLocations'
is set to NULL. Otherwise the 'rowTreeCuts' slot is set to NULL and
'rowCutLocations' is set to the given value. Similarly for 'col'.