Update vignette for non-centered demo data
The previously used TCGA.GBM demo data had values centered around 0, but
the new BRCA data does not. The difference affects the vignette for
continuous color maps and data layers, which had the color map
breakpoints explicitly set to -2, 0, 2.

This commit updates that vignette to account for this difference. This
difference is also used to show the utility of having different data
layers by making one layer the unadjusted data and row-centering the
second layer.