Actions bug fix: Reorder steps in pkgdown action
Fix the order of the steps in the pkgdown action so that the NGCHM
package is installed before rendering the index.Rmd file. This change is
needed because rendering the index.Rmd file requires the NGCHM package.

Some of the previous actions run from pkgdonw.yaml succeeded even with
this incorrect ordering because the NGCHM package was in the cache from
previous recent runs. (According to current GitHub documentation, caches
are removed after 7 days of disuse)