Replace explicit support for SingleCellExperiment/Seurat
Direct support for SingleCellExperiment/Seurat requires availability
of these packages.  This is unneeded overhead for users not in the
single-cell field.

This patch replaces direct support with a simple mechanism that enables
any package to extend the NGCHM package's capabilities in a similar way.
Our goal is to add that support directly into the Seurat and
SingleCellExperiment packages.