Don't run NGCHMDemoData/NGCHMSupportFiles examples
Because we don't want to have NGCHMDemoData nad NGCHMSupportFiles in
CRAN (mostly because CRAN doesn't want us to put these in CRAN), the
package must be able to build and run examples w/o them. One option is
to replace all the examples with generic matrix data....but I think it
is more useful to users to see the NGCHMDemoData. So opting for this
instead.