Add optional examples & cleanup indentation
We want to make the examples using the NGCHMDemoData only run if that
package is available. We also want other minimal examples if the
NGCHMDemoData is not available.

Additionally, I standardized the indentation in functions.R, and
performed a few other minor lintr suggestions. Ideally this would have
been a separate commit, but sometimes things are not ideal.