Make examples run / run faster
Some of the examples took longer than 5 seconds to run, so I made them
smaller. Some of the smaller examples using Rtsne and umap failed
because the default perplexity (Rtsne) or number of nearest neighbors
(umap) was to high. So I changed these parameters from the default for
the sake of having a small example that runs.