Run styler package on all R files
In an attempt to improve readability, which will help other developers
want to use our code, ran the styler package on all the R files and
regenerated the documentation.

I used the default settings for styler::style_file(), which means the
style guide is that of tidyverse.

I have not yet checked if these changes broke anything. Fixes for
anything broken will be in subsequent commits.