Do not write man page for non-exported functions
A number of internal functions have Roxygen2 documentation, but do not
document a return value. I'm also a little unsure if the existing
Roxygen2 documentation is complete. For these, adding @noRd so that a
man page will NOT be created.

For CRAN, it is acceptable to NOT have a man page for internal
functions.