Do not run examples that require git
Calls to `chmNew()` with data make calls to the internal function
`gitHashObject()`, which requires a system call to git. Git is not
available on the R-builder servers for Debian and Ubuntu (but seems to
be installed for Windows).

As such, any example that uses a call to `chmNew()` with matrix data
fails on these servers.

For reference, here is an example error message from R-hub builder
platforms "Debian Linux, R-release GCC" and "Debian Linux, R-devel GCC":

  sh: 1: git: not found
  Error in system2("git", c("hash-object", path), stdout = TRUE) :
    error in running command
  Calls: chmNew ... shaidyHashProtoBlob -> vapply -> FUN ->
           gitHashObject -> system2