Use tempdir() for default directories
The CRAN reviewer for 1.0.0 said:

  Please omit any default path in writing functions. In your
  examples/vignettes/tests you can write to tempdir().

So setting defaults for inpDir, outDir, and saveDir to tempdir().
Hopefully this is acceptable to the reviewer and does not ruin the user
experience.

FYI: I tried using NULL instead of tempdir(), but that caused errors in
simple usage because of the restriction that saveDir must be of type
'character'.