Add some helper classes for panel_config
These helper classes will be used to construct the panel_configuration
section of the mapConfig.json file as part of the feature request to
initialize the panel layout from this R package.

NOTE: these classes are a preliminary implementation and only handle the
default panel layout of two panes. This limitation is noted in the
comments with appropriate 'TODO' and 'FIX ME'.