Warn if missing external utilities / packages
Give the user a more informative message if missing any of the suggested
utilities (git, java 11, scp, ssh, tar) or packages (NGCHMSupportFiles,
NGCHMDemoData). Warning messages are formatted to facilitate
cut-and-paste.

CRAN does not allow us to install or load other packages from within our
package. So the best we can do is provide commands for the user.

Because ShaidyMapGen.jar is required to create .ngchm/.html files, AND
it is compiled with Java 11, the code explicitly checks for the java
version.

If needed, we can make the error messages for scp, ssh, and tar more
informative.

This addresses issue #46, #47, #49, and #50