Give warnings if missing needed support files
The functions `chmExportToFile()`, `chmExportToHTML()`, and
`chmExportToPDF()` all use ShaidyMapGen.jar, and `chmExportToHTML()`
also uses ngchmWidget-min.js. These functions allow the user to supply
the path, and if not supplied uses the appropriate environment variable.

Give the user better warning messages to let the user know if the paths
to these files are missing.

Addresses issue #52