Add git/java/NGCHMSupportFiles to functions docs
Add information about requiring git, java, and NGCHMSupportFiles to
functions that require these.

NOTE: the funky line breaks in the code block for installing
NGCHMSupportFiles are to avoid long-line truncation in the PDF manual.

NOTE: This does add more places were we hard-code the java version
requirement, which is unfortunate.

Addresses issue #48.