Run java headless in system2() calls
The existing syntax for setting the DISPLAY env (a 'fix' for issue #51)
is not correct and gives the following error on Windows only:

  Error: Could not find or load main class DISPLAY=''
  Caused by: java.lang.ClassNotFoundException: DISPLAY=''

Rather than fixing the syntax, we used a different approach to avoid
futzing with DISPLAY at all.

I manually tested this fix works on Windows, Linux (RedHat), MacOS, and
the ngchm/rstudio-ngchm:latest docker container.

Addresses issue #54.