Only build documentation on release
Now that we are in CRAN, we want the documentation website to be in
synch with the CRAN version. So instead of building the site with every
push to main or tag, only build site when there is a release.

This assumes we make a GitHub release only when we submit a new version
to CRAN.