Do not build site w/ every push to main
We don't want to update the documentation web site with every push to
main. In general, we want to keep the site in synch with the code in
CRAN. However we sometimes might want to update the documentation w/o
having a release.

So here changing the GitHub action for building the documentation
website to run automatically when there is a release, and also keep the
ability to run it manually (to allow for documentation updates in
between releases).

This addresses issue #57.