Move YouTube links to menu bar
At yesterday's meeting it was decided to remove the YouTube article
page, and replace it with a "Video Tutorials" menu dropdown.

By default, all the links open in the same tab--so clicking on one of
these menu items takes the user to a different site. I think that's
disorienting for users. So I added javascript to force external links to
open in a different tab. This does affect every external link on the
site, but I think that's OK.

For reference, the 'external-link' class is part of pkgdown API:
  https://pkgdown.r-lib.org/news/index.html
            ?q=external-link#html-css-and-js-2-0-0