Use stringsAsFactors=FALSE in vignettes
The default in R 4.x is stringsAsFactors = FALSE; however in R 3.x the
default is TRUE.

When stringsAsFactors = TRUE, the vignette examples fail because the
covariate values in the NGCHMDemoData ("WT" and "MUT") are read in as
factors, which are stored as integers (mode is "numeric"), causing
chmNewColorMap to error out trying to perform numerical operations on
character strings "WT" and "MUT".

Setting stringsAsFactors = FALSE may save users with R < 4.0 a little
frustration.

A more robust choice is to adjust this R package to better handle this
situation. However doing so may not bubble to the top of the priority
list. We'll discuss at the next meeting.