Update required R version to >= 3.5
There's a little bit of dependency :fire: here:

- NGCHM R package lists htmltools under Imports
- htmltools claims to require R >= 2.14.1, but
- htmltools imports rlang, which specifies R >= 3.5.0.

Because of these things, `install.packages("NGCHM")` fails on R 3.4.

Users can go into the archives of rlang and install version 1.0.0. Then,
provided they don't have compiler issues when fastmap installs (another
dependency of htmltools, I had to futz with my gcc versions to install
successfully), NGCHM will install and the vignettes will work in R 3.4.
I verified this with R 3.4.2 and gcc 7.0.0.

One other tiny wrinkle: the NGCHMDemoData (which of course is only
needed for the vignettes) requires 3.5 because of the formatting of the
.rda files:

https://github.com/MD-Anderson-Bioinformatics/NGCHMDemoData/commit/e8f1079

With all this, I think it's better to just say NGCHM requires R >= 3.5.
If users have a legacy system and need NGCHMs for an earlier version of
R, hopefully they'll reach out and we can help.