Explicitly set color of NG-CHM buttons
When the user switches between light and dark mode, elements such as
buttons are switched to the default light or dark mode colors unless
explicitly set.

Because we had the background color set, but not the text color, when
the user switched to dark mode, these icons were unreadable.

Since the purpose of these icons is to assist the user with
understanding the icons in the NG-CHM viewer, we probably want to set
them as the same colors as the NG-CHM viewer, and NOT let them change
when the user switches between light and dark mode.