Use more robust class checking in chmAddPCA
Making this change to avoid a NOTE from checking in preparation for
CRAN submission. For reference, here's the NOTE from running
`R CMD check --as-cran`:

* checking R code for possible problems ... NOTE
Found if() conditions comparing class() to string:
File ‘NGCHM/R/functions.R’:
  if (length(class(prc)) != 1 || class(prc) != "prcomp") ...
Use inherits() (or maybe is()) instead.

This is similar to what was done for the `hm` argument in 926f4dc.