Change warnings to messages in startup checks
Users have expressed frustrations with the warning messages from
the checks for NGCHMSupportFiles and NGCHMDemoData. To address these
frustrations:

- `warning()` changed to `packageStartUpMessage()`
  - Not using `message()` because that results in a NOTE about
    unsuppressable startup messages when building for CRAN.
- if NGCHMSupportFiles / NGCHMDemoData was installed but not loaded,
  the package now loads them instead of asking the user to load them.

Users have not mentioned the following, but these changes were made with
similar rationale:

In checking for system utilities (git, java 11, etc.), `warning()` was
changed to `packageStartUpMessage()`.

The language of the messages was also changed to (hopefully) improve
clarity.

The function documentation was also changed to (hopefully) help
developers. These functions are internal, so users never see their
documentation.