Include a last-ditch effort to get support files
In 399e228 we changed the functionality of checkForNGCHMSupportFiles()
to load NGCHMSupportFiles if possible. This means we can make things
slightly easier for the user with a last-ditch effort to get the needed
paths to the support files. The change in this commit does that for the
following scenario:

- User does not have NGCHMSupportFiles installed and has not set
  envs SHAIDYMAPGEN or NGCHMWIDGETPATH
- User tries to export an NGCHM to a file, but that fails because
  required support files are not available
- User sees message from checkForNGCHMSupportFiles, and installs
  but does not load NGCHMSupportFiles.

This last-ditch effort will then load the package and save the user from
having to load it manually.