Use packageStartupMessage in java/sys checks
Making this change to avoid the following NOTE when running `R CMD
CHECK` on systems w/o the suggested external utilities:

```
* checking whether startup messages can be suppressed ... NOTE
IMPORTANT: Unable to verify Java installation. Java 11 is required to
           create .ngchm and .html files.
      Install Java 11 or higher for to create .ngchm and .html files.

It looks like this package (or a package it requires) has a startup
message which cannot be suppressed: see ?packageStartupMessage.
```

This sort of NOTE has the potential to cause CRAN to reject the package.