Implement default panel layout
This implements a basic version of the feature request to initialize a
panel layout from R. This does not include all of the requested items in
issue #17. Indeed, the only new layout supported is a three-panel layout
with geometry:

          | pane2
    pane1 | ------
          | pane3

where the user can specify the content of each pane (e.g. pane1 is a
summary map, pane2 is a plugin, etc). The current defaul layout is also
supported:

   pane1 (detailMap) | pane2 (summaryMap)

NOTE: This functionality also depends on ShaidyMapGen.jar. The changes
in https://github.com/MD-Anderson-Bioinformatics/NG-CHM/pull/553 are
also required in order to produce a .ngchm file with panels initialized.

NOTE: This introduces unfortunate additional coupling to the NGCHM
viewer project. This is because of the lines like:

```javascript
  mapItem.dataBoxWidth = savedState.dataBoxWidth;
  ...
  let zoomBoxSizeIdx = DET.zoomBoxSizes.indexOf(mapItem.dataBoxWidth);
```

in DET.restoreFromSavedState (see DetailHeatMapDisplay.js). Because of
this, `dataBoxWidth` must both be defined and one of the
DET.zoomBoxSizes. I'm not sure there's a good way around this, given
that we don't want to introduce breaking changes into the NGCHM
ecosystem.