Remove restrictions on label display length
The rationale for these restrictions was the specified values in the
"Maximum Label Length" dropdown in the NG-CHM Viewer (to see this
dropdown: click main gear icon to open "Heat Map Display Properties" and
click on the "Rows & Cols" tab).

At yesterday's meeting, @bmbroom planned to change the "Maximum Label
Length" from a dropdown to a text box accepting integer values.
Anticipating this change, we decided to remove the restrictions here.